+++ bentoml/__main__.py
@@ -1,4 +1,4 @@
 if __name__ == "__main__":
-    from bentoml._internal.cli import cli
+    from bentoml_cli.cli import cli
 
     cli()
