+++ setup.cfg
@@ -88,7 +88,7 @@ include =
 
 [options.entry_points]
 console_scripts =
-    bentoml = bentoml_cli:cli
+    bentoml = bentoml_cli.cli:cli
 
 [options.extras_require]
 grpc =
